+++ CC/ml to cc/test.py
@@ -1,6 +1,6 @@
 import requests
 
-resp = requests.post("http://127.0.0.1:5000/",
+resp = requests.post("https://ricediseaseprediction-dumhlsbi3q-et.a.run.app",
                      files={'file': open('brownspot.jpg', 'rb')})
 
 print(resp.json())
